feat: add TransactionSuccessModal component to display transaction success details in SendFundsModal
